refactor to use event listeners to interact with videos on carousel change, update icons, move mute into cta buttons
