+++ site/static/js/carousel-video-mute-button.component.js
@@ -1,4 +1,4 @@
-import { AppComponent, bindAllComponents, render, h, attrs } from 's72.ui';
+import { AppComponent, bindAllComponents, render, h, attrs, Tooltip } from 's72.ui';
 import CarouselVideoMutedIcon from './icons/carousel-video-muted-icon.js';
 import CarouselVideoUnMutedIcon from './icons/carousel-video-unmuted-icon.js';
 
@@ -8,7 +8,9 @@ export default class CarouselVideoMuteButton extends AppComponent {
     this.state = { muted: true };
   }
 
-  toggleBackgroundMute() {
+  toggleBackgroundMute(e) {
+    e.preventDefault();
+
     let video = document.querySelector('.s72-carousel-item.current video');
 
     if (video.muted) {
@@ -22,9 +24,11 @@ export default class CarouselVideoMuteButton extends AppComponent {
 
   render() {
     return (
-      <button id="muteBtn" class="mute-btn" onClick={() => this.toggleBackgroundMute()}>
-        {this.state.muted ? <CarouselVideoMutedIcon /> : <CarouselVideoUnMutedIcon />}
-      </button>
+      <Tooltip tooltip={this.state.muted ? 'Unmute' : 'Mute'} position="bottom">
+        <button id="muteBtn" class="s72-btn mute-btn" onClick={e => this.toggleBackgroundMute(e)}>
+          {this.state.muted ? <CarouselVideoMutedIcon /> : <CarouselVideoUnMutedIcon />}
+        </button>
+      </Tooltip>
     );
   }
 }
