+++ site/styles/_carousel.scss
@@ -15,53 +15,6 @@ s72-carousel {
   }
 }
 
-.carousel-video-btns {
-  align-items: center;
-  bottom: 135px;
-  display: flex;
-  flex-direction: row;
-  height: 30px;
-  position: absolute;
-  right: 50px;
-  z-index: 99999;
-  @include media-breakpoint-up(xs) {
-    bottom: 142px;
-  }
-  @include media-breakpoint-up(lg) {
-    bottom: 18px;
-  }
-
-  .mute-btn {
-    background: transparent;
-    border: 0;
-    padding: 0;
-
-    svg {
-      fill: #fff;
-      height: 26px;
-      width: 26px;
-    }
-  }
-
-  .trailer-btn {
-    background-color: #fff;
-    border: 0;
-    height: 26px;
-    margin: 10px;
-    width: 26px;
-    /* stylelint-disable */
-    -webkit-mask-image: url('/images/icons/video-popup.png');
-    mask-image: url('/images/icons/video-popup.png');
-    -webkit-mask-size: 26px;
-    mask-size: 26px;
-    /* stylelint-enable */
-
-    .s72-btn-play-icon {
-      display: none;
-    }
-  }
-}
-
 .page-collection-carousel {
   display: flex;
   flex-direction: column;
